Updating README to set FINK_CLIENT_HOME global var
When attempting to run bin/fink_client_test.sh for the first time one
needs to ensure that $FINK_CLIENT_HOME is set beforehand. This will
ensure the variable to set.

However, there may be better way to ensure this env-var is set such that the user
does not need to set this themselves

	modified:   README.md